+++ ktor-network/ktor-network-test/build.gradle.kts
@@ -1,0 +1,3 @@
+/*
+ * Copyright 2014-2023 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
