+++ settings.gradle.kts
@@ -142,6 +142,7 @@ include(":ktor-io")
 include(":ktor-utils")
 include(":ktor-network")
 include(":ktor-network:ktor-network-tls")
+include(":ktor-network:ktor-network-test")
 include(":ktor-network:ktor-network-quic")
 include(":ktor-network:ktor-network-tls:ktor-network-tls-certificates")
 include(":ktor-bom")
